ml_service/requirements.txt
datasets==2.14.4
numpy==1.25.0
pandas==2.0.3
pika==1.3.2
pymongo==4.5.0
python-dotenv==1.0.0
scikit_learn==1.2.2
sentence_transformers==2.2.2
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
